feat[纹理]：点精灵

+++ sample/src/main/java/com/jiangpengyong/sample/utils/SizeUtils.kt
@@ -3,8 +3,8 @@ package com.jiangpengyong.sample.utils
 import android.content.res.Resources
 
 object SizeUtils {
-    fun dp2px(dpValue: Float): Int {
+    fun dp2px(dpValue: Float): Float {
         val scale = Resources.getSystem().displayMetrics.density
-        return (dpValue * scale + 0.5f).toInt()
+        return dpValue * scale + 0.5f
     }
 }
